requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
python-dotenv==1.0.0
aiohttp==3.9.1
pymongo==4.6.1
tenacity==8.2.3
psutil==5.9.6
Pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
